requirements.txt
asgiref==3.11.1
attrs==25.4.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==6.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.2
drf-spectacular==0.27.1
factory_boy==3.3.3
Faker==40.5.1
google-auth==2.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
packaging==26.0
pluggy==1.6.0
psycopg==3.1.18
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pytest-django==4.12.0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
